+++ core/src/test/java/kafka/admin/JConfigCommandTest.java
@@ -86,445 +86,9 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-public class ConfigCommandTest {
+public class JConfigCommandTest {
     private static final String ZK_CONNECT = "localhost:2181";
-    private static final DummyAdminZkClient DUMMY_ADMIN_ZK_CLIENT = new DummyAdminZkClient(null);
-
-    private static final List<String> ZOOKEEPER_BOOTSTRAP = Arrays.asList("--zookeeper", ZK_CONNECT);
-    private static final List<String> BROKER_BOOTSTRAP = Arrays.asList("--bootstrap-server", "localhost:9092");
-    private static final List<String> CONTROLLER_BOOTSTRAP = Arrays.asList("--bootstrap-controller", "localhost:9093");
-
-    @Test
-    public void shouldExitWithNonZeroStatusOnArgError() {
-        assertNonZeroStatusExit("--blah");
-    }
-
-    @Test
-    public void shouldExitWithNonZeroStatusOnZkCommandWithTopicsEntity() {
-        assertNonZeroStatusExit(toArray(ZOOKEEPER_BOOTSTRAP, Arrays.asList(
-            "--entity-type", "topics",
-            "--describe")));
-    }
-
-    @Test
-    public void shouldExitWithNonZeroStatusOnZkCommandWithClientsEntity() {
-        assertNonZeroStatusExit(toArray(ZOOKEEPER_BOOTSTRAP, Arrays.asList(
-            "--entity-type", "clients",
-            "--describe")));
-    }
-
-    @Test
-    public void shouldExitWithNonZeroStatusOnZkCommandWithIpsEntity() {
-        assertNonZeroStatusExit(toArray(ZOOKEEPER_BOOTSTRAP, Arrays.asList(
-            "--entity-type", "ips",
-            "--describe")));
-    }
-
-    @Test
-    public void shouldExitWithNonZeroStatusAlterUserQuotaWithoutEntityName() {
-        assertNonZeroStatusExit(toArray(BROKER_BOOTSTRAP, Arrays.asList(
-            "--entity-type", "users",
-            "--alter", "--add-config", "consumer_byte_rate=20000")));
-    }
-
-    @Test
-    public void shouldExitWithNonZeroStatusOnBrokerCommandError() {
-        assertNonZeroStatusExit("--bootstrap-server", "invalid host",
-            "--entity-type", "brokers",
-            "--entity-name", "1",
-            "--describe");
-    }
-
-    @Test
-    public void shouldExitWithNonZeroStatusIfBothBootstrapServerAndBootstrapControllerGiven() {
-        assertNonZeroStatusExit(toArray(BROKER_BOOTSTRAP, CONTROLLER_BOOTSTRAP, Arrays.asList(
-            "--describe", "--broker-defaults")));
-    }
-
-    @Test
-    public void shouldExitWithNonZeroStatusOnBrokerCommandWithZkTlsConfigFile() {
-        assertNonZeroStatusExit(
-            "--bootstrap-server", "invalid host",
-            "--entity-type", "users",
-            "--zk-tls-config-file", "zk_tls_config.properties",
-            "--describe");
-    }
-
-    @Test
-    public void shouldFailParseArgumentsForClientsEntityTypeUsingZookeeper() {
-        assertThrows(IllegalArgumentException.class, () -> testArgumentParse(ZOOKEEPER_BOOTSTRAP, "clients"));
-    }
-
-    @Test
-    public void shouldParseArgumentsForClientsEntityTypeWithBrokerBootstrap() {
-        testArgumentParse(BROKER_BOOTSTRAP, "clients");
-    }
-
-    @Test
-    public void shouldParseArgumentsForClientsEntityTypeWithControllerBootstrap() {
-        testArgumentParse(CONTROLLER_BOOTSTRAP, "clients");
-    }
-
-    @Test
-    public void shouldParseArgumentsForUsersEntityTypeUsingZookeeper() {
-        testArgumentParse(ZOOKEEPER_BOOTSTRAP, "users");
-    }
-
-    @Test
-    public void shouldParseArgumentsForUsersEntityTypeWithBrokerBootstrap() {
-        testArgumentParse(BROKER_BOOTSTRAP, "users");
-    }
-
-    @Test
-    public void shouldParseArgumentsForUsersEntityTypeWithControllerBootstrap() {
-        testArgumentParse(CONTROLLER_BOOTSTRAP, "users");
-    }
-
-    @Test
-    public void shouldFailParseArgumentsForTopicsEntityTypeUsingZookeeper() {
-        assertThrows(IllegalArgumentException.class, () -> testArgumentParse(ZOOKEEPER_BOOTSTRAP, "topics"));
-    }
-
-    @Test
-    public void shouldParseArgumentsForTopicsEntityTypeWithBrokerBootstrap() {
-        testArgumentParse(BROKER_BOOTSTRAP, "topics");
-    }
-
-    @Test
-    public void shouldParseArgumentsForTopicsEntityTypeWithControllerBootstrap() {
-        testArgumentParse(CONTROLLER_BOOTSTRAP, "topics");
-    }
-
-    @Test
-    public void shouldParseArgumentsForBrokersEntityTypeUsingZookeeper() {
-        testArgumentParse(ZOOKEEPER_BOOTSTRAP, "brokers");
-    }
-
-    @Test
-    public void shouldParseArgumentsForBrokersEntityTypeWithBrokerBootstrap() {
-        testArgumentParse(BROKER_BOOTSTRAP, "brokers");
-    }
-
-    @Test
-    public void shouldParseArgumentsForBrokersEntityTypeWithControllerBootstrap() {
-        testArgumentParse(CONTROLLER_BOOTSTRAP, "brokers");
-    }
-
-    @Test
-    public void shouldParseArgumentsForBrokerLoggersEntityTypeWithBrokerBootstrap() {
-        testArgumentParse(BROKER_BOOTSTRAP, "broker-loggers");
-    }
-
-    @Test
-    public void shouldParseArgumentsForBrokerLoggersEntityTypeWithControllerBootstrap() {
-        testArgumentParse(CONTROLLER_BOOTSTRAP, "broker-loggers");
-    }
-
-    @Test
-    public void shouldFailParseArgumentsForIpEntityTypeUsingZookeeper() {
-        assertThrows(IllegalArgumentException.class, () -> testArgumentParse(ZOOKEEPER_BOOTSTRAP, "ips"));
-    }
-
-    @Test
-    public void shouldParseArgumentsForIpEntityTypeWithBrokerBootstrap() {
-        testArgumentParse(BROKER_BOOTSTRAP, "ips");
-    }
-
-    @Test
-    public void shouldParseArgumentsForIpEntityTypeWithControllerBootstrap() {
-        testArgumentParse(CONTROLLER_BOOTSTRAP, "ips");
-    }
-
-    public void testArgumentParse(List<String> bootstrapArguments, String entityType) {
-        String shortFlag = "--" + entityType.substring(0, entityType.length() - 1);
-        String connectOpts1 = bootstrapArguments.get(0);
-        String connectOpts2 = bootstrapArguments.get(1);
-
-        // Should parse correctly
-        ConfigCommandOptions createOpts = new ConfigCommandOptions(toArray(connectOpts1, connectOpts2,
-            "--entity-name", "1",
-            "--entity-type", entityType,
-            "--describe"));
-        createOpts.checkArgs();
-
-        createOpts = new ConfigCommandOptions(toArray(connectOpts1, connectOpts2,
-            shortFlag, "1",
-            "--describe"));
-        createOpts.checkArgs();
-
-        // For --alter and added config
-        createOpts = new ConfigCommandOptions(toArray(connectOpts1, connectOpts2,
-            "--entity-name", "1",
-            "--entity-type", entityType,
-            "--alter",
-            "--add-config", "a=b,c=d"));
-        createOpts.checkArgs();
-
-        createOpts = new ConfigCommandOptions(toArray(connectOpts1, connectOpts2,
-            "--entity-name", "1",
-            "--entity-type", entityType,
-            "--alter",
-            "--add-config-file", "/tmp/new.properties"));
-        createOpts.checkArgs();
-
-        createOpts = new ConfigCommandOptions(toArray(connectOpts1, connectOpts2,
-            shortFlag, "1",
-            "--alter",
-            "--add-config", "a=b,c=d"));
-        createOpts.checkArgs();
-
-        createOpts = new ConfigCommandOptions(toArray(connectOpts1, connectOpts2,
-            shortFlag, "1",
-            "--alter",
-            "--add-config-file", "/tmp/new.properties"));
-        createOpts.checkArgs();
-
-        // For alter and deleted config
-        createOpts = new ConfigCommandOptions(toArray(connectOpts1, connectOpts2,
-            "--entity-name", "1",
-            "--entity-type", entityType,
-            "--alter",
-            "--delete-config", "a,b,c"));
-        createOpts.checkArgs();
-
-        createOpts = new ConfigCommandOptions(toArray(connectOpts1, connectOpts2,
-            shortFlag, "1",
-            "--alter",
-            "--delete-config", "a,b,c"));
-        createOpts.checkArgs();
-
-        // For alter and both added, deleted config
-        createOpts = new ConfigCommandOptions(toArray(connectOpts1, connectOpts2,
-            "--entity-name", "1",
-            "--entity-type", entityType,
-            "--alter",
-            "--add-config", "a=b,c=d",
-            "--delete-config", "a"));
-        createOpts.checkArgs();
-
-        createOpts = new ConfigCommandOptions(toArray(connectOpts1, connectOpts2,
-            shortFlag, "1",
-            "--alter",
-            "--add-config", "a=b,c=d",
-            "--delete-config", "a"));
-        createOpts.checkArgs();
-
-        Properties addedProps = ConfigCommand.parseConfigsToBeAdded(createOpts);
-        assertEquals(2, addedProps.size());
-        assertEquals("b", addedProps.getProperty("a"));
-        assertEquals("d", addedProps.getProperty("c"));
-
-        Seq<String> deletedProps = ConfigCommand.parseConfigsToBeDeleted(createOpts);
-        assertEquals(1, deletedProps.size());
-        assertEquals("a", deletedProps.apply(0));
-
-        createOpts = new ConfigCommandOptions(toArray(connectOpts1, connectOpts2,
-            "--entity-name", "1",
-            "--entity-type", entityType,
-            "--alter",
-            "--add-config", "a=b,c=,d=e,f="));
-        createOpts.checkArgs();
-
-        createOpts = new ConfigCommandOptions(toArray(connectOpts1, connectOpts2,
-            shortFlag, "1",
-            "--alter",
-            "--add-config", "a._-c=b,c=,d=e,f="));
-        createOpts.checkArgs();
-
-        Properties addedProps2 = ConfigCommand.parseConfigsToBeAdded(createOpts);
-        assertEquals(4, addedProps2.size());
-        assertEquals("b", addedProps2.getProperty("a._-c"));
-        assertEquals("e", addedProps2.getProperty("d"));
-        assertTrue(addedProps2.getProperty("c").isEmpty());
-        assertTrue(addedProps2.getProperty("f").isEmpty());
-
-        ConfigCommandOptions inValidCreateOpts = new ConfigCommandOptions(toArray(connectOpts1, connectOpts2,
-            shortFlag, "1",
-            "--alter",
-            "--add-config", "a;c=b"));
-
-        assertThrows(IllegalArgumentException.class,
-            () -> ConfigCommand.parseConfigsToBeAdded(inValidCreateOpts));
-
-        ConfigCommandOptions inValidCreateOpts2 = new ConfigCommandOptions(toArray(connectOpts1, connectOpts2,
-            shortFlag, "1",
-            "--alter",
-            "--add-config", "a,=b"));
-
-        assertThrows(IllegalArgumentException.class,
-            () -> ConfigCommand.parseConfigsToBeAdded(inValidCreateOpts2));
-    }
-
-    @Test
-    public void shouldFailIfAddAndAddFile() {
-        // Should not parse correctly
-        ConfigCommandOptions createOpts = new ConfigCommandOptions(toArray("--bootstrap-server", "localhost:9092",
-            "--entity-name", "1",
-            "--entity-type", "brokers",
-            "--alter",
-            "--add-config", "a=b,c=d",
-            "--add-config-file", "/tmp/new.properties"
-        ));
-        assertThrows(IllegalArgumentException.class, createOpts::checkArgs);
-    }
-
-    @Test
-    public void testParseConfigsToBeAddedForAddConfigFile() {
-        String fileContents =
-            "a=b\n" +
-            "c = d\n" +
-            "json = {\"key\": \"val\"}\n" +
-            "nested = [[1, 2], [3, 4]]";
-
-        File file = TestUtils.tempFile(fileContents);
-
-        List<String> addConfigFileArgs = Arrays.asList("--add-config-file", file.getPath());
-
-        ConfigCommandOptions createOpts = new ConfigCommandOptions(toArray(Arrays.asList("--bootstrap-server", "localhost:9092",
-                "--entity-name", "1",
-                "--entity-type", "brokers",
-                "--alter"),
-            addConfigFileArgs));
-        createOpts.checkArgs();
-
-        Properties addedProps = ConfigCommand.parseConfigsToBeAdded(createOpts);
-        assertEquals(4, addedProps.size());
-        assertEquals("b", addedProps.getProperty("a"));
-        assertEquals("d", addedProps.getProperty("c"));
-        assertEquals("{\"key\": \"val\"}", addedProps.getProperty("json"));
-        assertEquals("[[1, 2], [3, 4]]", addedProps.getProperty("nested"));
-    }
-
-    @SuppressWarnings("deprecation") // Added for Scala 2.12 compatibility for usages of JavaConverters
-    public void testExpectedEntityTypeNames(List<String> expectedTypes, List<String> expectedNames, List<String> connectOpts, String...args) {
-        ConfigCommandOptions createOpts = new ConfigCommandOptions(toArray(Arrays.asList(connectOpts.get(0), connectOpts.get(1), "--describe"), Arrays.asList(args)));
-        createOpts.checkArgs();
-        assertEquals(createOpts.entityTypes().toSeq(), ConfigCommandIntegrationTest.seq(expectedTypes));
-        assertEquals(createOpts.entityNames().toSeq(), ConfigCommandIntegrationTest.seq(expectedNames));
-    }
-
-    public void doTestOptionEntityTypeNames(boolean zkConfig) {
-        List<String> connectOpts = zkConfig
-            ? Arrays.asList("--zookeeper", ZK_CONNECT)
-            : Arrays.asList("--bootstrap-server", "localhost:9092");
-
-        // zookeeper config only supports "users" and "brokers" entity type
-        if (!zkConfig) {
-            testExpectedEntityTypeNames(Collections.singletonList(ConfigType.TOPIC), Collections.singletonList("A"), connectOpts, "--entity-type", "topics", "--entity-name", "A");
-            testExpectedEntityTypeNames(Collections.singletonList(ConfigType.IP), Collections.singletonList("1.2.3.4"), connectOpts, "--entity-name", "1.2.3.4", "--entity-type", "ips");
-            testExpectedEntityTypeNames(Arrays.asList(ConfigType.USER, ConfigType.CLIENT), Arrays.asList("A", ""), connectOpts,
-                "--entity-type", "users", "--entity-type", "clients", "--entity-name", "A", "--entity-default");
-            testExpectedEntityTypeNames(Arrays.asList(ConfigType.USER, ConfigType.CLIENT), Arrays.asList("", "B"), connectOpts,
-                "--entity-default", "--entity-name", "B", "--entity-type", "users", "--entity-type", "clients");
-            testExpectedEntityTypeNames(Collections.singletonList(ConfigType.TOPIC), Collections.singletonList("A"), connectOpts, "--topic", "A");
-            testExpectedEntityTypeNames(Collections.singletonList(ConfigType.IP), Collections.singletonList("1.2.3.4"), connectOpts, "--ip", "1.2.3.4");
-            testExpectedEntityTypeNames(Arrays.asList(ConfigType.CLIENT, ConfigType.USER), Arrays.asList("B", "A"), connectOpts, "--client", "B", "--user", "A");
-            testExpectedEntityTypeNames(Arrays.asList(ConfigType.CLIENT, ConfigType.USER), Arrays.asList("B", ""), connectOpts, "--client", "B", "--user-defaults");
-            testExpectedEntityTypeNames(Arrays.asList(ConfigType.CLIENT, ConfigType.USER), Collections.singletonList("A"), connectOpts,
-                "--entity-type", "clients", "--entity-type", "users", "--entity-name", "A");
-            testExpectedEntityTypeNames(Collections.singletonList(ConfigType.TOPIC), Collections.emptyList(), connectOpts, "--entity-type", "topics");
-            testExpectedEntityTypeNames(Collections.singletonList(ConfigType.IP), Collections.emptyList(), connectOpts, "--entity-type", "ips");
-        }
-
-        testExpectedEntityTypeNames(Collections.singletonList(ConfigType.BROKER), Collections.singletonList("0"), connectOpts, "--entity-name", "0", "--entity-type", "brokers");
-        testExpectedEntityTypeNames(Collections.singletonList(ConfigType.BROKER), Collections.singletonList("0"), connectOpts, "--broker", "0");
-        testExpectedEntityTypeNames(Collections.singletonList(ConfigType.USER), Collections.emptyList(), connectOpts, "--entity-type", "users");
-        testExpectedEntityTypeNames(Collections.singletonList(ConfigType.BROKER), Collections.emptyList(), connectOpts, "--entity-type", "brokers");
-    }
-
-    @Test
-    public void testOptionEntityTypeNamesUsingZookeeper() {
-        doTestOptionEntityTypeNames(true);
-    }
-
-    @Test
-    public void testOptionEntityTypeNames() {
-        doTestOptionEntityTypeNames(false);
-    }
-
-    @Test
-    public void shouldFailIfUnrecognisedEntityTypeUsingZookeeper() {
-        ConfigCommandOptions createOpts = new ConfigCommandOptions(new String[]{"--zookeeper", ZK_CONNECT,
-            "--entity-name", "client", "--entity-type", "not-recognised", "--alter", "--add-config", "a=b,c=d"});
-        assertThrows(IllegalArgumentException.class, () -> ConfigCommand.alterConfigWithZk(null, createOpts, DUMMY_ADMIN_ZK_CLIENT));
-    }
-
-    @Test
-    public void shouldFailIfUnrecognisedEntityType() {
-        ConfigCommandOptions createOpts = new ConfigCommandOptions(new String[]{"--bootstrap-server", "localhost:9092",
-            "--entity-name", "client", "--entity-type", "not-recognised", "--alter", "--add-config", "a=b,c=d"});
-        assertThrows(IllegalArgumentException.class, () -> ConfigCommand.alterConfig(new DummyAdminClient(new Node(1, "localhost", 9092)), createOpts));
-    }
-
-    @Test
-    public void shouldFailIfBrokerEntityTypeIsNotAnIntegerUsingZookeeper() {
-        ConfigCommandOptions createOpts = new ConfigCommandOptions(new String[]{"--zookeeper", ZK_CONNECT,
-            "--entity-name", "A", "--entity-type", "brokers", "--alter", "--add-config", "a=b,c=d"});
-        assertThrows(IllegalArgumentException.class, () -> ConfigCommand.alterConfigWithZk(null, createOpts, DUMMY_ADMIN_ZK_CLIENT));
-    }
-
-    @Test
-    public void shouldFailIfBrokerEntityTypeIsNotAnInteger() {
-        ConfigCommandOptions createOpts = new ConfigCommandOptions(new String[]{"--bootstrap-server", "localhost:9092",
-            "--entity-name", "A", "--entity-type", "brokers", "--alter", "--add-config", "a=b,c=d"});
-        assertThrows(IllegalArgumentException.class, () -> ConfigCommand.alterConfig(new DummyAdminClient(new Node(1, "localhost", 9092)), createOpts));
-    }
-
-    @Test
-    public void shouldFailIfShortBrokerEntityTypeIsNotAnIntegerUsingZookeeper() {
-        ConfigCommandOptions createOpts = new ConfigCommandOptions(new String[]{"--zookeeper", ZK_CONNECT,
-            "--broker", "A", "--alter", "--add-config", "a=b,c=d"});
-        assertThrows(IllegalArgumentException.class, () -> ConfigCommand.alterConfigWithZk(null, createOpts, DUMMY_ADMIN_ZK_CLIENT));
-    }
-
-    @Test
-    public void shouldFailIfShortBrokerEntityTypeIsNotAnInteger() {
-        ConfigCommandOptions createOpts = new ConfigCommandOptions(new String[]{"--bootstrap-server", "localhost:9092",
-            "--broker", "A", "--alter", "--add-config", "a=b,c=d"});
-        assertThrows(IllegalArgumentException.class, () -> ConfigCommand.alterConfig(new DummyAdminClient(new Node(1, "localhost", 9092)), createOpts));
-    }
-
-    @Test
-    public void shouldFailIfMixedEntityTypeFlagsUsingZookeeper() {
-        ConfigCommandOptions createOpts = new ConfigCommandOptions(new String[]{"--zookeeper", ZK_CONNECT,
-            "--entity-name", "A", "--entity-type", "users", "--client", "B", "--describe"});
-        assertThrows(IllegalArgumentException.class, createOpts::checkArgs);
-    }
-
-    @Test
-    public void shouldFailIfMixedEntityTypeFlags() {
-        ConfigCommandOptions createOpts = new ConfigCommandOptions(new String[]{"--bootstrap-server", "localhost:9092",
-            "--entity-name", "A", "--entity-type", "users", "--client", "B", "--describe"});
-        assertThrows(IllegalArgumentException.class, createOpts::checkArgs);
-    }
-
-    @Test
-    public void shouldFailIfInvalidHost() {
-        ConfigCommandOptions createOpts = new ConfigCommandOptions(new String[]{"--bootstrap-server", "localhost:9092",
-            "--entity-name", "A,B", "--entity-type", "ips", "--describe"});
-        assertThrows(IllegalArgumentException.class, createOpts::checkArgs);
-    }
-
-    @Test
-    public void shouldFailIfInvalidHostUsingZookeeper() {
-        ConfigCommandOptions createOpts = new ConfigCommandOptions(new String[]{"--zookeeper", ZK_CONNECT,
-            "--entity-name", "A,B", "--entity-type", "ips", "--describe"});
-        assertThrows(IllegalArgumentException.class, createOpts::checkArgs);
-    }
-
-    @Test
-    public void shouldFailIfUnresolvableHost() {
-        ConfigCommandOptions createOpts = new ConfigCommandOptions(new String[]{"--bootstrap-server", "localhost:9092",
-            "--entity-name", "RFC2606.invalid", "--entity-type", "ips", "--describe"});
-        assertThrows(IllegalArgumentException.class, createOpts::checkArgs);
-    }
-
-    @Test
-    public void shouldFailIfUnresolvableHostUsingZookeeper() {
-        ConfigCommandOptions createOpts = new ConfigCommandOptions(new String[]{"--zookeeper", ZK_CONNECT,
-            "--entity-name", "RFC2606.invalid", "--entity-type", "ips", "--describe"});
-        assertThrows(IllegalArgumentException.class, createOpts::checkArgs);
-    }
+    private static final ConfigCommandUnitTest.DummyAdminZkClient DUMMY_ADMIN_ZK_CLIENT = new ConfigCommandUnitTest.DummyAdminZkClient(null);
 
     @Test
     public void shouldAddClientConfigUsingZookeeper() {
@@ -1920,33 +1484,6 @@ public class ConfigCommandTest {
             .flatMap(Collection::stream)
             .forEach(e -> res.put(e.getKey(), e.getValue()));
         return res;
-    }
-
-    static class DummyAdminZkClient extends AdminZkClient {
-        public DummyAdminZkClient(KafkaZkClient zkClient) {
-            super(zkClient, scala.None$.empty());
-        }
-
-        @Override
-        public void changeBrokerConfig(Seq<Object> brokers, Properties configs) {
-        }
-
-        @Override
-        public Properties fetchEntityConfig(String rootEntityType, String sanitizedEntityName) {
-            return new Properties();
-        }
-
-        @Override
-        public void changeClientIdConfig(String sanitizedClientId, Properties configs) {
-        }
-
-        @Override
-        public void changeUserOrUserClientIdConfig(String sanitizedEntityName, Properties configs, boolean isUserClientId) {
-        }
-
-        @Override
-        public void changeTopicConfig(String topic, Properties configs) {
-        }
     }
 
     static class DummyAdminClient extends MockAdminClient {
