+++ core/src/test/scala/unit/kafka/admin/ConfigCommandTest.scala
@@ -45,90 +45,6 @@ class ConfigCommandTest extends Logging {
   private val dummyAdminZkClient = new DummyAdminZkClient(null)
 
   @Test
-  def shouldFailIfUnrecognisedEntityTypeUsingZookeeper(): Unit = {
-    val createOpts = new ConfigCommandOptions(Array("--zookeeper", zkConnect,
-      "--entity-name", "client", "--entity-type", "not-recognised", "--alter", "--add-config", "a=b,c=d"))
-    assertThrows(classOf[IllegalArgumentException], () => ConfigCommand.alterConfigWithZk(null, createOpts, dummyAdminZkClient))
-  }
-
-  @Test
-  def shouldFailIfUnrecognisedEntityType(): Unit = {
-    val createOpts = new ConfigCommandOptions(Array("--bootstrap-server", "localhost:9092",
-      "--entity-name", "client", "--entity-type", "not-recognised", "--alter", "--add-config", "a=b,c=d"))
-    assertThrows(classOf[IllegalArgumentException], () => ConfigCommand.alterConfig(new DummyAdminClient(new Node(1, "localhost", 9092)), createOpts))
-  }
-
-  @Test
-  def shouldFailIfBrokerEntityTypeIsNotAnIntegerUsingZookeeper(): Unit = {
-    val createOpts = new ConfigCommandOptions(Array("--zookeeper", zkConnect,
-      "--entity-name", "A", "--entity-type", "brokers", "--alter", "--add-config", "a=b,c=d"))
-    assertThrows(classOf[IllegalArgumentException], () => ConfigCommand.alterConfigWithZk(null, createOpts, dummyAdminZkClient))
-  }
-
-  @Test
-  def shouldFailIfBrokerEntityTypeIsNotAnInteger(): Unit = {
-    val createOpts = new ConfigCommandOptions(Array("--bootstrap-server", "localhost:9092",
-      "--entity-name", "A", "--entity-type", "brokers", "--alter", "--add-config", "a=b,c=d"))
-    assertThrows(classOf[IllegalArgumentException], () => ConfigCommand.alterConfig(new DummyAdminClient(new Node(1, "localhost", 9092)), createOpts))
-  }
-
-  @Test
-  def shouldFailIfShortBrokerEntityTypeIsNotAnIntegerUsingZookeeper(): Unit = {
-    val createOpts = new ConfigCommandOptions(Array("--zookeeper", zkConnect,
-      "--broker", "A", "--alter", "--add-config", "a=b,c=d"))
-    assertThrows(classOf[IllegalArgumentException], () => ConfigCommand.alterConfigWithZk(null, createOpts, dummyAdminZkClient))
-  }
-
-  @Test
-  def shouldFailIfShortBrokerEntityTypeIsNotAnInteger(): Unit = {
-    val createOpts = new ConfigCommandOptions(Array("--bootstrap-server", "localhost:9092",
-      "--broker", "A", "--alter", "--add-config", "a=b,c=d"))
-    assertThrows(classOf[IllegalArgumentException], () => ConfigCommand.alterConfig(new DummyAdminClient(new Node(1, "localhost", 9092)), createOpts))
-  }
-
-  @Test
-  def shouldFailIfMixedEntityTypeFlagsUsingZookeeper(): Unit = {
-    val createOpts = new ConfigCommandOptions(Array("--zookeeper", zkConnect,
-      "--entity-name", "A", "--entity-type", "users", "--client", "B", "--describe"))
-    assertThrows(classOf[IllegalArgumentException], () => createOpts.checkArgs())
-  }
-
-  @Test
-  def shouldFailIfMixedEntityTypeFlags(): Unit = {
-    val createOpts = new ConfigCommandOptions(Array("--bootstrap-server", "localhost:9092",
-      "--entity-name", "A", "--entity-type", "users", "--client", "B", "--describe"))
-    assertThrows(classOf[IllegalArgumentException], () => createOpts.checkArgs())
-  }
-
-  @Test
-  def shouldFailIfInvalidHost(): Unit = {
-    val createOpts = new ConfigCommandOptions(Array("--bootstrap-server", "localhost:9092",
-      "--entity-name", "A,B", "--entity-type", "ips", "--describe"))
-    assertThrows(classOf[IllegalArgumentException], () => createOpts.checkArgs())
-  }
-
-  @Test
-  def shouldFailIfInvalidHostUsingZookeeper(): Unit = {
-    val createOpts = new ConfigCommandOptions(Array("--zookeeper", zkConnect,
-      "--entity-name", "A,B", "--entity-type", "ips", "--describe"))
-    assertThrows(classOf[IllegalArgumentException], () => createOpts.checkArgs())
-  }
-
-  @Test
-  def shouldFailIfUnresolvableHost(): Unit = {
-    val createOpts = new ConfigCommandOptions(Array("--bootstrap-server", "localhost:9092",
-      "--entity-name", "RFC2606.invalid", "--entity-type", "ips", "--describe"))
-    assertThrows(classOf[IllegalArgumentException], () => createOpts.checkArgs())
-  }
-
-  @Test
-  def shouldFailIfUnresolvableHostUsingZookeeper(): Unit = {
-    val createOpts = new ConfigCommandOptions(Array("--zookeeper", zkConnect,
-      "--entity-name", "RFC2606.invalid", "--entity-type", "ips", "--describe"))
-    assertThrows(classOf[IllegalArgumentException], () => createOpts.checkArgs())
-  }
-
-  @Test
   def shouldAddClientConfigUsingZookeeper(): Unit = {
     val createOpts = new ConfigCommandOptions(Array("--zookeeper", zkConnect,
       "--entity-name", "my-client-id",
